Add support for building with 64-bit architectures on OS X.
This patch brings full 64-bit compatibility to Synergy on OS X by
replacing all obsolete 32-bit only Carbon functions with modern
equivalents. All functions introduced have been available since 10.4 so
this won't affect the minimum deployment target. Specifically:

Creating an empty CGEvent and getting its location is behaviourally
identical to GetGlobalMouse, and yes, both are in flipped coordinates.
This was tested with a multi-monitor configuration as well.

TrackMouseLocationWithOptions is behaviourally identical to
GetGlobalMouse in these cases because the timeout was 0 and none of the
other out params were used, except for the MouseTrackingResult in one
call was checked against kMouseTrackingTimedOut. Since the timeout was 0
and not kEventDurationForever, that value never could have been returned
anyway.

Instead of attempting to define SIntXX and UIntXX manually, MacTypes.h
is included on OS X. These types were wrong in 64-bit mode because of
this, causing type redefinition errors.

+++ src/lib/synergy/DaemonApp.cpp
@@ -302,9 +302,9 @@ CDaemonApp::handleIpcMessage(const CEvent& e, void*)
 				LOG((CLOG_DEBUG "new command, elevate=%d command=%s", cm->elevate(), command.c_str()));
 
 				CString debugArg("--debug");
-				UInt32 debugArgPos = static_cast<UInt32>(command.find(debugArg));
+				size_t debugArgPos = command.find(debugArg);
 				if (debugArgPos != CString::npos) {
-					UInt32 from = debugArgPos + static_cast<UInt32>(debugArg.size()) + 1;
+					UInt32 from = static_cast<UInt32>(debugArgPos) + static_cast<UInt32>(debugArg.size()) + 1;
 					UInt32 nextSpace = static_cast<UInt32>(command.find(" ", from));
 					CString logLevel(command.substr(from, nextSpace - from));
